base_station_anomaly_detection/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
dash>=2.0.0
haversine>=2.5.0
geopy>=2.2.0
folium>=0.12.0
tqdm>=4.62.0
pyyaml>=5.4.0
joblib>=1.1.0
xgboost>=1.5.0
lightgbm>=3.3.0
streamlit>=1.10.0